+++ backend/dump.py
@@ -1,0 +1,2 @@
+games={}  # {game_id: Connect4}
+users={}  # {user_id: User}
